Separate Menu and Sidebar components
