+++ src/mui/layout/Layout.js
@@ -9,6 +9,7 @@ import withWidth from 'material-ui/utils/withWidth';
 import compose from 'recompose/compose';
 import injectTapEventPlugin from 'react-tap-event-plugin';
 import AppBar from './AppBar';
+import Sidebar from './Sidebar';
 import Notification from './Notification';
 import defaultTheme from '../defaultTheme';
 import { setSidebarVisibility as setSidebarVisibilityAction } from '../../actions';
@@ -61,7 +62,6 @@ class Layout extends Component {
             isLoading,
             menu,
             route,
-            sidebarOpen,
             theme,
             title,
             width,
@@ -80,10 +80,9 @@ class Layout extends Component {
                     { width !== 1 && <AppBar title={title} />}
                     <div className="body" style={width === 1 ? prefixedStyles.bodySmall : prefixedStyles.body}>
                         <div style={width === 1 ? prefixedStyles.contentSmall : prefixedStyles.content}>{children}</div>
-                        {React.cloneElement(menu, {
-                            open: sidebarOpen,
-                            width,
-                        })}
+                        <Sidebar>
+                            {menu}
+                        </Sidebar>
                     </div>
                     <Notification />
                     {isLoading && <CircularProgress
@@ -105,7 +104,6 @@ Layout.propTypes = {
     menu: PropTypes.element,
     route: PropTypes.object.isRequired,
     setSidebarVisibility: PropTypes.func.isRequired,
-    sidebarOpen: PropTypes.bool,
     title: PropTypes.string.isRequired,
     theme: PropTypes.object.isRequired,
     width: PropTypes.number,
@@ -118,7 +116,6 @@ Layout.defaultProps = {
 function mapStateToProps(state) {
     return {
         isLoading: state.admin.loading > 0,
-        sidebarOpen: state.admin.ui.sidebarOpen,
     };
 }
